daemon: extract secrets-backend module from config.rs
Continues #41. Pulls all secret-store backend I/O — the file store,
systemd credentials, libsecret/keyring (via `secret-tool` with a
keyring-rs fallback), and KWallet (via `kwallet-query`) — into
`config/secrets.rs`. Schema-side helpers (`SecretConfig`,
`default_secret_account`, `resolve_secret_account`,
`common_secret_file_path`, `default_secret_store_dir`,
`default_secret_service`, `default_wallet_*`, `resolve_wallet_entry`)
stay in `mod.rs` because they're also used by setting setters and
views unrelated to backend I/O. The submodule reaches them via
`super::…`.

Public surface kept stable via `use secrets::{…}` re-imports in
`mod.rs` so existing call sites (`set_api_key`'s redacted-audit
logging, settings setters that gate on placeholder values, the
parent test module's exhaustive coverage of the helper functions)
keep working without prefix changes. The single test-only helper
(`sanitize_secret_value`) is referenced as `secrets::…` from the test
module to avoid a cfg-gated `use`.

`jwt.rs` had two callers of `read_common_file_secret` /
`write_common_file_secret` (the JWT signing key persistence layer)
which were already routing through `super::…`; updated to point at
`super::secrets::…` for the new location.

Drops the orphaned `std::process::{Command, Stdio}` and
`keyring::Entry` imports from `mod.rs` (they were only used by the
backends).

`config/mod.rs`: 3861 → 3536 lines. New `config/secrets.rs` lands at
375 lines. Net split this PR: ~330 lines moved into a self-contained
module.

Cumulative across the four #41 PRs: `config/mod.rs` 4373 → 3536
(-837 lines), now organized as `mod.rs` plus four focused
submodules (`pam_auth`, `jwt`, `oidc`, `secrets`).
